rename fhir endpoint file
